Extracted logic about ScalingFactor/vertical speed from RedDotsData into its own class - VerticalSpeed
